✅ adjusted test with pre-trained models triggered only via GitHub CI
